style: bar 제거 및 ErrorInfo 컴포넌트 분리
